added validation to delete thread

+++ app/Http/Controllers/ThreadsController.php
@@ -122,9 +122,11 @@ class ThreadsController extends Controller
      */
     public function destroy(Channel $channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
-        if (\request()->wantsJson()) {
+        if (request()->wantsJson()) {
             return response([], 204);
         }
